wp_id on uint8

+++ sw/airborne/modules/computer_vision/tag_tracking.c
@@ -179,7 +179,7 @@ struct tag_info {
 };
 
 struct wp_tracking {
-  uint16_t wp_id;
+  uint8_t wp_id;
   int16_t tag_id;
 };
